# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MédiGuide API",
    "description": "API de localisation de professionnels de santé (médecins, pharmacies, parapharmacies) à Jendouba, Tunisie : recherche géographique, prise de rendez-vous, avis, espace professionnel et administration. Un module complémentaire (medieye) expose ces mêmes fiches au format Facility générique de la plateforme partagée mediEYE/craftEYE/cityEYE sous /api/modules/medieye — voir openapi-medieye-fragment.json.",
    "version": "1.0.0",
    "contact": {
      "name": "MédiGuide"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Développement local"
    },
    {
      "url": "https://<votre-projet>.up.railway.app/api",
      "description": "Production (Railway)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT d'accès posé en cookie httpOnly après /api/auth/login"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Requis sur toute requête POST/PATCH/DELETE ; valeur du cookie csrf_token obtenu via GET /api/auth/csrf"
      }
    },
    "schemas": {
...
      "Facility": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nom": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "medecin",
              "pharmacie",
              "parapharmacie"
            ]
          },
          "specialite": {
            "type": "string",
            "nullable": true
          },
          "adresse": {
            "type": "string"
          },
          "delegation": {
            "type": "string"
          },
          "telephone": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "noteAvis": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "distanceMetres": {
            "type": "number",
            "nullable": true,
            "description": "Présent uniquement si lat/lng fournis dans la recherche"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/facilities": {
      "get": {
        "tags": [
          "Facilities"
        ],
        "summary": "Rechercher médecins / pharmacies / parapharmacies",
        "security": [],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "medecin",
                "pharmacie",
                "parapharmacie"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Recherche full-text nom + spécialité"
          },
          {
            "name": "specialite",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ville",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "typeGarde",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 5000
            },
            "description": "Rayon en mètres, plafonné à 50000"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 300
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des établissements (résultat mis en cache Redis 60s si géo)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Facility"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Catégorie invalide"
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Lister les utilisateurs",
        "responses": {
          "200": {
            "description": "Liste"
          },
          "403": {
            "description": "Rôle administrateur requis"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Créer un utilisateur",
        "responses": {
          "201": {
            "description": "Créé"
          }
        }
      }
    },
    "/admin/users/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Modifier un utilisateur",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modifié"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Supprimer un utilisateur",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Supprimé"
          }
        }
      }
    },
...
    "/admin/medecins/{id}/validation": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Valider / invalider une fiche médecin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statut mis à jour"
          }
        }
      }
    },
...
    "/admin/avis/{id}/moderation": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Modérer un avis (masquer/rétablir)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statut mis à jour"
          }
        }
      }
    },
...
  }
}